Schedule Demo Wireframe
This commit is to add functionality to the "Schedule a Demo" button on the contact section of the main page. Currently, this is only a wireframe with functionality needing to be added later. I added the demo dialog component which will prompt users to select a day (currently is set up to only include the current date). The user will then be prompted to select a 30 minute block of time between 9am and 4:30pm. Then the user will enter his/her name, organization, email, and description of what they would like to see from the demo. Then will then be able to submit. Currently, the submit is unfunctional, therefore it just clears the popup.