Trace and Expand Table Buttons Implementation
Added new UX buttons to the tables which will allow users to more easily navigate to the Trace page for individual requirements, as well as view the popup. The buttons are functioning all on different tables, and require the user to select a cell to see the menu with the Trace and Expand buttons. These buttons are not visible while the user is in edit mode.