Gemini Mermaid Prototype + React Change
Been cooking this for a while, finally pushing. Took me a while to push cause I wanted to test it but FastAPI has been down for a while and I don't want to wait any longer to push so I'm pushing with potential errors. NOTE: The errors, as far as I can tell, are purely prompt-based. So I don't think this will blow up, but keep caution when using it if you wanna test it before the final version is pushed. More info in the file.

TLDR: Adding mermaid Gemini functionality to fetch_req_diagramType.py, made a tiny change to the related GildeEditableTable.tsx file (added peopleNames to the end of a useCallback call within the GildeEditableTable function. Do Ctrl+F and search for Evan to see the change on line 910). Also did npx next lint --fix and npx prettier --write . so a bunch of other files were changed.